+++ WPF-Miao/Jian/search/UserController.cs
@@ -36,7 +36,7 @@ namespace Jian.search
                 }
                 var root = response.JsonBody.RootElement;
 
-                var code = root.GetProperty("resultCode").GetInt16();
+                var code = root.GetProperty("resultCode").GetInt32();
                 var success = root.GetProperty("success").GetBoolean();
                 if (code != 20000 || !success)
                 {
